requirements.txt
viam-sdk==0.52.0
numpy>=1.21.0
typing-extensions